+++ lib/DhanHQ/rate_limiter.rb
@@ -140,24 +140,58 @@ module DhanHQ
 
     # Spawns background threads to reset counters after each interval elapses.
     def start_cleanup_threads
+      @cleanup_threads = []
+      @shutdown = Concurrent::AtomicBoolean.new(false)
+
       # Don't create per_second cleanup thread - we handle it with timestamps
-      Thread.new do
+      @cleanup_threads << Thread.new do
         loop do
+          break if @shutdown.true?
+
           sleep(60)
-          @buckets[:per_minute]&.value = 0
+          break if @shutdown.true?
+
+          mutex.synchronize do
+            @buckets[:per_minute]&.value = 0
+          end
         end
       end
-      Thread.new do
+
+      @cleanup_threads << Thread.new do
         loop do
+          break if @shutdown.true?
+
           sleep(3600)
-          @buckets[:per_hour]&.value = 0
+          break if @shutdown.true?
+
+          mutex.synchronize do
+            @buckets[:per_hour]&.value = 0
+          end
         end
       end
-      Thread.new do
+
+      @cleanup_threads << Thread.new do
         loop do
+          break if @shutdown.true?
+
           sleep(86_400)
-          @buckets[:per_day]&.value = 0
+          break if @shutdown.true?
+
+          mutex.synchronize do
+            @buckets[:per_day]&.value = 0
+          end
         end
+      end
+    end
+
+    # Shuts down cleanup threads gracefully
+    def shutdown
+      return if @shutdown.true?
+
+      @shutdown.make_true
+      @cleanup_threads&.each do |thread|
+        thread&.wakeup if thread&.alive?
+        thread&.join(5) # Wait up to 5 seconds for thread to finish
       end
     end
   end